move isLabeledPoi to constants.ts
